+++ src/Core/Managed/Shared/DataContracts/SessionStateTelemetry.cs
@@ -2,25 +2,24 @@
 {
     using System;
     using Microsoft.ApplicationInsights.Channel;
-    using Microsoft.ApplicationInsights.Extensibility.Implementation.External;
 
     /// <summary>
     /// Telemetry type used to track user sessions.
     /// </summary>
+    [Obsolete("Session state events are no longer used. This telemetry item will be sent as EventTelemetry.")]
     public sealed class SessionStateTelemetry : ITelemetry
     {
-        internal const string TelemetryName = "SessionState";
+        internal readonly EventTelemetry Data;
 
-        internal readonly SessionStateData Data;
-        private readonly TelemetryContext context;
-        
+        private readonly string startEventName = "Session started";
+        private readonly string endEventName = "Session ended";
+
         /// <summary>
         /// Initializes a new instance of the <see cref="SessionStateTelemetry"/> class.
         /// </summary>
         public SessionStateTelemetry()
+            : this(SessionState.Start)
         {
-            this.Data = new SessionStateData();
-            this.context = new TelemetryContext();
         }
 
         /// <summary>
@@ -29,33 +28,81 @@
         /// <param name="state">
         /// A <see cref="SessionState"/> value indicating state of the user session.
         /// </param>
-        public SessionStateTelemetry(SessionState state) : this()
+        public SessionStateTelemetry(SessionState state)
         {
+            this.Data = new EventTelemetry();
             this.State = state;
         }
 
         /// <summary>
         /// Gets or sets the date and time the session state was recorded.
         /// </summary>
-        public DateTimeOffset Timestamp { get; set; }
+        public DateTimeOffset Timestamp
+        {
+            get
+            {
+                return this.Data.Timestamp;
+            }
+
+            set
+            {
+                this.Data.Timestamp = value;
+            }
+        }
 
         /// <summary>
         /// Gets the <see cref="TelemetryContext"/> of the application when the session state was recorded.
         /// </summary>
         public TelemetryContext Context
         {
-            get { return this.context; }
+            get { return this.Data.Context; }
         }
 
         /// <summary>
         /// Gets or sets the value that defines absolute order of the telemetry item.
         /// </summary>
-        public string Sequence { get; set; }
+        public string Sequence
+        {
+            get
+            {
+                return this.Data.Sequence;
+            }
+
+            set
+            {
+                this.Data.Sequence = value;
+            }
+        }
 
         /// <summary>
         /// Gets or sets the value describing state of the user session.
         /// </summary>
-        public SessionState State { get; set; }
+        public SessionState State
+        {
+            get
+            {
+                if (this.Data.Name == this.startEventName)
+                {
+                    return SessionState.Start;
+                }
+                else
+                {
+                    return SessionState.End;
+                }
+            }
+
+            set
+            {
+                if (value == SessionState.Start)
+                {
+                    this.Data.Name = this.startEventName;
+                }
+                else
+                {
+                    this.Data.Name = this.endEventName;
+                }
+            }
+        }
 
         /// <summary>
         /// Sanitizes this telemetry instance to ensure it can be accepted by the Application Insights.